+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## [0.5.0] - 2025-11-27
+
+Features:
+- Dedicated UI for 2 players
+- User preferences for disabling UI features
+- Card grouping in play
+- Alignment guides in play
+- Highlight cards in hands usable in the current phase
+- Animate card lock/unlock
+
+Bugfix:
+- Last oust now gives 2 VP
+- Cards dragged on lib/crypt now always go on top
+- Prevent infinite loop on the bot
+
 ## [0.4.1] - 2025-11-17
 
 Features: